Fall back to text mutator on USD attribute type mismatch
RealityKit's USD reader is type-strict on RealityKitComponent fields:
HierarchicalFade.opacity is authored by RCP as `float`, and the runtime
silently skips a `double opacity` value (sphere stays fully opaque even
when 0.25 is in the file). That's the actual cause of "opacity doesn't
visually update" that I'd been hand-waving away as a viewport caching
issue. It is not.

Two coordinated changes:

1. setComponentParameterWithUSDMutation: when attr.Set returns false
   (which happens when the requested VtValue type doesn't match the
   attribute's authored type — e.g. Float into an existing Double
   attribute, or vice versa), return false instead of throwing
   componentAuthoringFailed. The caller already falls through to the
   USDA text mutator on a `false` return, which rewrites the attribute
   type prefix in the layer. That gives us the orphan's behavior:
   inspector decides the type, USDA gets that type, RealityKit reads
   the type it expects.

2. ComponentParameterEditor.scalar reverts to "float". The orphan
   componentParameterAuthoringSpec table (InspectorFeature.swift:3833)
   declares HierarchicalFade.opacity as float; that table reflects what
   RealityKit's RCP-compatible reader actually wants. Same goes for
   AmbientAudio.gain, SpatialAudio fields, etc. — all the scalar
   component params are written as float.

After this, the first opacity edit on a previously-authored `double
opacity = X` line rewrites it to `float opacity = X`. RealityKit then
applies the value on reload and the sphere renders translucent. Pre-
existing scenes whose opacity was authored as `double` need one edit
to migrate; new components author as `float` from the start.

+++ Packages/DeconstructedLibrary/Sources/InspectorShellUI/InspectorView.swift
@@ -272,7 +272,7 @@ private struct ComponentParameterEditor: View {
 			LabeledContent(unit.map { "\(parameter.label) (\($0))" } ?? parameter.label) {
 				TextField("", value: Binding(
 					get: { value },
-					set: { onChange("double", String($0)) }
+					set: { onChange("float", String($0)) }
 				), format: .number.precision(.fractionLength(0...4)))
 				.textFieldStyle(.roundedBorder)
 				.frame(minWidth: 80)
